requirements_txt.txt
Flask==2.3.3
groq==0.4.1
python-dotenv==1.0.0
Werkzeug==2.3.7